refactor: extract quality alerts logic into reusable QualityAlerts component
